+++ gDRplots/R/plot-pheatmap.R
@@ -168,7 +168,7 @@ pheatmap_qc <- function(
   names(drug_annotation_colors) <- drug_to_colored
   
   # dendrogram
-  cluster_rows <- if (cluster_rows && !any(is.na(mat_cvd))) {
+  cluster_rows <- if (cluster_rows && !any(is.na(mat_cvd)) && NROW(mat_cvd) >= 2) {
     stats::hclust(stats::dist(mat_cvd))
   } else {
     FALSE
@@ -230,7 +230,7 @@ pheatmap_qc <- function(
 }
 
 
-#' Plot pretty heatmap with annotationsfor single-agent data
+#' Plot pretty heatmap with annotations for single-agent data
 #'
 #' @param dt_metrics \code{data.table} representing data from the \code{Metrics} assay,
 #'  outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
@@ -277,11 +277,11 @@ pheatmap_qc <- function(
 #' ggpubr::as_ggplot(hm_1[["gtable"]])
 #' 
 #' annotation_manual_col <-
-#'   unique(dt_metrics[, .SD, .SDcols = c("CellLineName", "Tissue")])
+#'   unique(dt_metrics[, c("CellLineName", "Tissue"), with = FALSE])
 #' annotation_manual_row <-
-#'   unique(dt_metrics[, .SD, .SDcols = c("DrugName", "drug_moa")])
+#'   unique(dt_metrics[, c("DrugName", "drug_moa"), with = FALSE])
 #' annotation_map <-
-#'   get_ann_color_map(unique(dt_metrics[,.SD, .SDcols = c("Tissue", "drug_moa")]))
+#'   get_ann_color_map(unique(dt_metrics[, c("Tissue", "drug_moa"), with = FALSE]))
 #' 
 #' output <- pheatmap_with_anno_sa(dt_metrics = dt_metrics,
 #'                                 normalization_type = "RV",
@@ -342,6 +342,9 @@ pheatmap_with_anno_sa <- function(
     annotation_col = NULL,
     annotation_colors = NULL) {
   
+  cellline_name <- gDRutils::get_env_identifiers("cellline_name")
+  drug_name <- gDRutils::get_env_identifiers("drug_name")
+  
   checkmate::assert_data_table(dt_metrics)
   checkmate::assert_choice(normalization_type, choices = c("GR", "RV"))
   numeric_columns <- names(dt_metrics)[vapply(dt_metrics, is.numeric, logical(1))]
@@ -354,11 +357,14 @@ pheatmap_with_anno_sa <- function(
   checkmate::assert_flag(cluster_rows)
   checkmate::assert_flag(cluster_cols)
   checkmate::assert_data_table(annotation_row, null.ok = TRUE)
+  if (!is.null(annotation_row)) {
+    checkmate::assert_names(names(annotation_row), must.include = drug_name)
+  }
   checkmate::assert_data_table(annotation_col, null.ok = TRUE)
+  if (!is.null(annotation_col)) {
+    checkmate::assert_names(names(annotation_col), must.include = cellline_name)
+  }
   checkmate::assert_list(annotation_colors, null.ok = TRUE)
-  
-  cellline_name <- gDRutils::get_env_identifiers("cellline_name")
-  drug_name <- gDRutils::get_env_identifiers("drug_name")
   
   # output
   ls_output <- list(data = list(matrix = NULL,
@@ -452,7 +458,7 @@ pheatmap_with_anno_sa <- function(
   
   # dendrogram
   cluster_condition <- !any(is.na(t_mat_cvd)) && !any(is.infinite(t_mat_cvd)) && 
-    NROW(t_mat_cvd) * NCOL(t_mat_cvd) <= 200 # gDR standard
+    any(dim(t_mat_cvd) < 200) # gDR standard
   cluster_rows <- if (cluster_rows && cluster_condition && NROW(t_mat_cvd) >= 2) {
     stats::hclust(stats::dist(t_mat_cvd))
   } else {
@@ -499,6 +505,296 @@ pheatmap_with_anno_sa <- function(
   return(ls_output)
 }
 
+#' Plot pretty heatmap with annotations for co-dilution data
+#'
+#' @param dt_metrics \code{data.table} representing data from the \code{Metrics} assay,
+#'  outputted by \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
+#'  and co-dilution \code{SummarizedExperiment}
+#' @param normalization_type string with normalization types to be selected
+#'                           one of: "GR" ("GRvalue") or "RV" ("RelativeViability")
+#' @param metric string name of metric;
+#'  one of: "xc50" ("GR50" or "IC50" - respectively depending on \code{normalization_type}), 
+#'  "x_max" ("GR Max" or "E Max") or "x_mean" ("GR Mean" or "RV Mean")
+#' @param fit_source string source name for metrics
+#' @param hm_title string plot title
+#' @param colors_vec character vector of colors (valid name or hex) used in heatmap;
+#'   note that the first color will be assigned to the min value, and the last one - to the max
+#' @param no_breaks numeric number of breaks on scale
+#' @param annotation_row \code{data.table} that specifies the annotations shown on the left side 
+#'   of the heatmap.
+#'   Each row defines the features for a specific row. The rows in the data and in the annotation
+#'   are matched using corresponding names from the required  \code{DrugName}, 
+#'   \code{DrugName_2} and \code{Concentration_2} columns.
+#'   Note that color schemes takes into account if the variable is continuous or discrete.
+#' @param cluster_rows logical flag whether indicating rows should be clustered;
+#'   the dendrogram will not be shown for the matrix with any NA values
+#' @param cluster_cols logical flag indicating whether columns should be clustered;
+#'   the dendrogram will not be shown for the matrix with any NA values or -Inf/Inf value
+#' @param annotation_col \code{data.table} that specifies the annotations shown above the heatmap.
+#'   Each row defines the features for a specific column. The columns in the data and in the annotation
+#'   are matched using corresponding names from the required  \code{CellLineName} column.
+#'   Note that color schemes takes into account if the variable is continuous or discrete.
+#' @param annotation_colors named list for specifying \code{annotation_col} and \code{annotation_row} 
+#'   track colors manually; note list is named with annotation name (column names of \code{annotation_row} - 
+#'   without \code{DrugName} and column names of \code{annotation_col} - 
+#'   without \code{CellLineName}), each list item is named vector with valid color name for 
+#'   each value described in \code{annotation_row} and in \code{annotation_col}, respectively.
+#'   Not described elements will be colored by default.
+#' 
+#' @seealso \code{\link[pheatmap]{pheatmap}}
+#'
+#' @examples
+#' mae <- gDRutils::get_synthetic_data("combo_codilution_small")
+#' se <- mae[[gDRutils::get_supported_experiments("cd")]]
+#' dt_metrics <- gDRutils::convert_se_assay_to_dt(se = se,
+#'                                                assay_name = "Metrics")
+#' 
+#' output <- pheatmap_with_anno_cd(dt_metrics = dt_metrics)
+#' hm_1 <- output[["heatmap"]]
+#' ggpubr::as_ggplot(hm_1[["gtable"]])
+#' 
+#' annotation_manual_col <-
+#'   unique(dt_metrics[, c("CellLineName", "Tissue"), with = FALSE])
+#' annotation_manual_row <-
+#'   unique(dt_metrics[, c("DrugName", "DrugName_2", "Concentration_2",
+#'                         "drug_moa", "drug_moa_2"),
+#'                     with = FALSE])
+#' annotation_map <-
+#'   get_ann_color_map(unique(dt_metrics[, c("Tissue", "drug_moa", "drug_moa_2"), with = FALSE]))
+#' 
+#' output <- pheatmap_with_anno_cd(dt_metrics = dt_metrics,
+#'                                 normalization_type = "RV",
+#'                                 metric = "x_mean",
+#'                                 colors_vec = c("darkblue", "grey90"),
+#'                                 annotation_row = annotation_manual_row,
+#'                                 annotation_col = annotation_manual_col,
+#'                                 annotation_colors = annotation_map)
+#' hm_2 <- output[["heatmap"]]
+#' ggpubr::as_ggplot(hm_2[["gtable"]])
+#' 
+#' annotation_manual <- data.table::data.table(
+#'   CellLineName =
+#'     c("cellline_AA", "cellline_EA", "cellline_IB", "cellline_MC", "cellline_BC"),
+#'   mut_A = c(1, 1, 1, 0, 0),
+#'   mut_B = c("yes", "yes", "no", "no", "no")
+#' )
+#' annotation_map <- list(
+#'   mut_A = c("1" = "coral", "0" = "cadetblue"),
+#'   mut_B = c("yes" = "black", "no" = "grey90")
+#' )
+#' 
+#' output <- pheatmap_with_anno_cd(dt_metrics = dt_metrics,
+#'                                 annotation_col = annotation_manual,
+#'                                 annotation_colors = annotation_map,
+#'                                 hm_title = get_hm_title(
+#'                                   normalization_type = "GR",
+#'                                   metric = "hsa_score",
+#'                                   dataset_name = "Co-dilution data"))
+#' hm_3 <- output[["heatmap"]]
+#' ggpubr::as_ggplot(hm_3[["gtable"]])
+#' 
+#' @keywords pheat_ann
+#' 
+#' @return A named list with elements:
+#' \itemize{
+#'   \item \code{data} a list containing the information visualized in the heatmap:
+#'     \itemize{
+#'       \item \code{matrix} data shown in the heatmap for the selected metric.
+#'       \item \code{annotation_row} a table with row annotations (for \code{DrugName}), if provided.
+#'       \item \code{annotation_col} a table with column annotations (for \code{CellLineName}), if provided.
+#'     }
+#'   \item \code{heatmap} the heatmap itself.
+#' }
+#' 
+#' @export
+pheatmap_with_anno_cd <- function(
+    dt_metrics,
+    normalization_type = "GR",
+    metric = "xc50",
+    fit_source = "gDR",
+    hm_title = NA,
+    colors_vec = c("firebrick2", "white"),
+    no_breaks = 50,
+    cluster_rows = TRUE,
+    cluster_cols = TRUE,
+    annotation_row = NULL,
+    annotation_col = NULL,
+    annotation_colors = NULL) {
+  
+  cellline_name <- gDRutils::get_env_identifiers("cellline_name")
+  drug_name <- gDRutils::get_env_identifiers("drug_name")
+  drug_name_2 <- gDRutils::get_env_identifiers("drug_name2")
+  conc_2 <- gDRutils::get_env_identifiers("concentration2")
+  
+  checkmate::assert_data_table(dt_metrics)
+  checkmate::assert_choice(normalization_type, choices = c("GR", "RV"))
+  numeric_columns <- names(dt_metrics)[vapply(dt_metrics, is.numeric, logical(1))]
+  checkmate::assert_choice(metric, choices = numeric_columns)
+  checkmate::assert_string(fit_source, null.ok = TRUE)
+  checkmate::assert_string(hm_title, na.ok = TRUE)
+  checkmate::assert_character(colors_vec)
+  stopifnot("Must be a valid color name" = all(vapply(colors_vec, is_valid_color, logical(1))))
+  checkmate::assert_int(no_breaks, lower = 2)
+  checkmate::assert_flag(cluster_rows)
+  checkmate::assert_flag(cluster_cols)
+  checkmate::assert_data_table(annotation_row, null.ok = TRUE)
+  if (!is.null(annotation_row)) {
+    checkmate::assert_names(names(annotation_row), must.include = c(drug_name, drug_name_2, conc_2))
+  }
+  checkmate::assert_data_table(annotation_col, null.ok = TRUE)
+  if (!is.null(annotation_col)) {
+    checkmate::assert_names(names(annotation_col), must.include = cellline_name)
+  }
+  checkmate::assert_list(annotation_colors, null.ok = TRUE)
+  
+  # output
+  ls_output <- list(data = list(matrix = NULL,
+                                annotation_col = NULL,
+                                annotation_row = NULL),
+                    heatmap = NULL)
+  
+  # select data for normalization type
+  filter_expr <- substitute(normalization_type == norm_type & fit_source == fit_src,
+                            list(norm_type = normalization_type, fit_src = fit_source))
+  tab_response <- dt_metrics[eval(filter_expr)]
+  
+  qmfun <- switch(metric,
+                  "xc50" = log10,
+                  identity)
+  
+  # prep data
+  tab_plot <- data.table::dcast(
+    data = tab_response,
+    formula = get(cellline_name) ~ paste(get(drug_name), "x", paste0(get(drug_name_2), "__", get(conc_2))),
+    value.var = metric)
+  data.table::setnames(tab_plot, "cellline_name", cellline_name)
+  
+  # prep matrix
+  mat_cvd <- as.matrix(tab_plot[, .SD, .SDcols = -cellline_name])
+  rownames(mat_cvd) <- tab_plot[[cellline_name]]
+  rm_col <- vapply(colnames(mat_cvd), function(i) !all(is.na(mat_cvd[, i])), logical(1))
+  rm_row <- vapply(seq_along(rownames(mat_cvd)), function(i) !all(is.na(mat_cvd[i, ])), logical(1))
+  if (!all(rm_col)) mat_cvd <- mat_cvd[, rm_col]
+  if (!all(rm_row)) mat_cvd <- mat_cvd[rm_row, ]
+  
+  # check completeness of annotation - TODO wrap in separate function
+  if (!is.null(annotation_col)) {
+    if (!all(rownames(mat_cvd) %in% annotation_col[[cellline_name]])) {
+      tab_missing_ann <- data.table::data.table(
+        missing = rownames(mat_cvd)[!rownames(mat_cvd) %in% annotation_col[[cellline_name]]]
+      )
+      data.table::setnames(tab_missing_ann, "missing", cellline_name)
+      
+      annotation_col <- data.table::rbindlist(list(annotation_col, tab_missing_ann), fill = TRUE)
+    }
+    # data.table::nafill does not support character
+    cols <- names(annotation_col)[names(annotation_col) != cellline_name]
+    data.table::setorderv(annotation_col, cols = cols, na.last = TRUE)
+    annotation_col[, (cols) := lapply(.SD, change_NA_into_char, "NA"), .SDcols = cols]
+    # select annotation acc to matrix
+    annotation_col <- annotation_col[get(cellline_name) %in% rownames(mat_cvd), ]
+    ls_output[["data"]][["annotation_col"]] <- annotation_col
+    
+    rownames(annotation_col) <- annotation_col[[cellline_name]] # required by pheatmap::pheatmap
+    annotation_col <- annotation_col[, .SD, .SDcol = -cellline_name]
+    # order matrix
+    mat_cvd <- mat_cvd[rownames(annotation_col), , drop = FALSE]
+  }
+  
+  if (!is.null(annotation_row)) {
+    DrugCombination <- NULL # due to NSE notes in R CMD check
+    annotation_row$DrugCombination <-
+      paste(annotation_row[[drug_name]], "x", paste0(annotation_row[[drug_name_2]], "__", annotation_row[[conc_2]]))
+    
+    if (!all(colnames(mat_cvd) %in% annotation_row[["DrugCombination"]])) {
+      tab_missing_ann <- data.table::data.table(
+        missing = colnames(mat_cvd)[!colnames(mat_cvd) %in% annotation_row[["DrugCombination"]]]
+      )
+      data.table::setnames(tab_missing_ann, "missing", "DrugCombination")
+      tab_missing_ann[, c(drug_name, drug_name_2) := data.table::tstrsplit(DrugCombination, " x ", fixed = TRUE)]
+      tab_missing_ann[, c(drug_name_2, conc_2) := data.table::tstrsplit(get(drug_name_2), "__", fixed = TRUE)]
+      tab_missing_ann[[conc_2]] <- as.numeric(tab_missing_ann[[conc_2]])
+      
+      annotation_row <- data.table::rbindlist(list(annotation_row, tab_missing_ann), fill = TRUE)
+    }
+    # data.table::nafill does not support character
+    cols <- names(annotation_row)[!names(annotation_row) %in% c(drug_name, drug_name_2, conc_2, "DrugCombination")]
+    data.table::setorderv(annotation_row, cols = cols, na.last = TRUE)
+    annotation_row[, (cols) := lapply(.SD, change_NA_into_char, "NA"), .SDcols = cols, drop = FALSE]
+    # select annotation acc to matrix
+    annotation_row <- annotation_row[DrugCombination %in% colnames(mat_cvd), ]
+    ls_output[["data"]][["annotation_row"]] <- annotation_row[, !c("DrugCombination"), with = FALSE]
+    
+    rownames(annotation_row) <- annotation_row[["DrugCombination"]] # required by pheatmap::pheatmap
+    annotation_row <- 
+      annotation_row[, .SD, .SDcols = -c(drug_name, drug_name_2, conc_2, "DrugCombination")]
+    # order matrix
+    mat_cvd <- mat_cvd[, rownames(annotation_row), drop = FALSE]
+  }
+  
+  # filling missing values
+  if (!is.null(annotation_col) && !is.null(annotation_colors)) {
+    annotation_colors <- fill_ann_color_map(annotation_col, annotation_colors)
+  }
+  if (!is.null(annotation_row) && !is.null(annotation_colors)) {
+    annotation_colors <- fill_ann_color_map(annotation_row, annotation_colors)
+  }
+  
+  ls_output[["data"]][["matrix"]] <- data.table::as.data.table(mat_cvd, keep.rownames = cellline_name)
+  # flip
+  t_mat_cvd <- t(mat_cvd)
+  t_mat_cvd[] <- vapply(t_mat_cvd, function(x) purrr::quietly(qmfun)(x)$result, numeric(1))
+  
+  # dendrogram
+  cluster_condition <- !any(is.na(t_mat_cvd)) && !any(is.infinite(t_mat_cvd)) && 
+    any(dim(t_mat_cvd) < 200) # gDR standard
+  cluster_rows <- if (cluster_rows && cluster_condition && NROW(t_mat_cvd) >= 2) {
+    stats::hclust(stats::dist(t_mat_cvd))
+  } else {
+    FALSE
+  }
+  cluster_cols <- if (cluster_cols && cluster_condition && NCOL(t_mat_cvd) >= 2) {
+    stats::hclust(stats::dist(t(t_mat_cvd)))
+  } else {
+    FALSE
+  }
+  
+  # prep hm color palette
+  min_val <- min(t_mat_cvd, na.rm = TRUE)
+  max_val <- ifelse(metric %in% c("x", "xc50", "x_max", "x_mean"), 1.0, max(t_mat_cvd, na.rm = TRUE))
+  
+  if (min_val == max_val) {
+    min_val <- min_val - 1
+  } else if (is.infinite(min_val)) {
+    min_val <- max_val / 100
+  }
+  
+  breaks <- seq(from = min_val, to = max_val, length.out = no_breaks)
+  hm_color_palette <- grDevices::colorRampPalette(colors_vec)(no_breaks + 1)
+  
+  # display numbers - for readability, turn it off for matrices larger than 10x10
+  display_numbers_flag <- !any(dim(t_mat_cvd) > c(10, 10))
+  
+  ls_output[["heatmap"]] <- 
+    pheatmap::pheatmap(mat = t_mat_cvd,
+                       scale = "none",
+                       display_numbers = display_numbers_flag,
+                       number_color = "black",
+                       color = hm_color_palette,
+                       breaks = breaks,
+                       angle_col = 90,
+                       main = hm_title,
+                       # dendrogram
+                       cluster_rows = cluster_rows,
+                       cluster_cols = cluster_cols,
+                       # manual annotation
+                       annotation_row = annotation_row,
+                       annotation_col = annotation_col,
+                       annotation_colors = annotation_colors,
+                       silent = TRUE)
+  
+  return(ls_output)
+}
 
 #' Plot pretty heatmap with annotations for combo data
 #' 
@@ -529,11 +825,11 @@ pheatmap_with_anno_sa <- function(
 #' dt_scores <- gDRutils::convert_se_assay_to_dt(se = se,
 #'                                               assay_name = "scores")
 #' annotation_manual_col <-
-#'   unique(dt_scores[, .SD, .SDcols = c("CellLineName", "Tissue")])
+#'   unique(dt_scores[, c("CellLineName", "Tissue"), with = FALSE])
 #' annotation_manual_row <-
-#'   unique(dt_scores[, .SD, .SDcols = c("DrugName", "DrugName_2", "drug_moa", "drug_moa_2")])
+#'   unique(dt_scores[, c("DrugName", "DrugName_2", "drug_moa", "drug_moa_2"), with = FALSE])
 #' annotation_map <-
-#'   get_ann_color_map(unique(dt_scores[,.SD, .SDcols = c("Tissue", "drug_moa", "drug_moa_2")]))
+#'   get_ann_color_map(unique(dt_scores[, c("Tissue", "drug_moa", "drug_moa_2"), with = FALSE]))
 #' 
 #' output <- pheatmap_with_anno_combo(dt_scores = dt_scores,
 #'                                    normalization_type = "RV",
@@ -596,6 +892,10 @@ pheatmap_with_anno_combo <- function(
     annotation_col = NULL,
     annotation_colors = NULL) {
   
+  cellline_name <- gDRutils::get_env_identifiers("cellline_name")
+  drug_name <- gDRutils::get_env_identifiers("drug_name")
+  drug_name_2 <- gDRutils::get_env_identifiers("drug_name2")
+  
   checkmate::assert_data_table(dt_scores)
   checkmate::assert_choice(normalization_type, choices = c("GR", "RV"))
   checkmate::assert_choice(metric, choices = c("hsa_score", "bliss_score"))
@@ -607,12 +907,14 @@ pheatmap_with_anno_combo <- function(
   checkmate::assert_flag(cluster_rows)
   checkmate::assert_flag(cluster_cols)
   checkmate::assert_data_table(annotation_row, null.ok = TRUE)
+  if (!is.null(annotation_row)) {
+    checkmate::assert_names(names(annotation_row), must.include = c(drug_name, drug_name_2))
+  }
   checkmate::assert_data_table(annotation_col, null.ok = TRUE)
+  if (!is.null(annotation_col)) {
+    checkmate::assert_names(names(annotation_col), must.include = cellline_name)
+  }
   checkmate::assert_list(annotation_colors, null.ok = TRUE)
-  
-  cellline_name <- gDRutils::get_env_identifiers("cellline_name")
-  drug_name <- gDRutils::get_env_identifiers("drug_name")
-  drug_name_2 <- gDRutils::get_env_identifiers("drug_name2")
   
   # output
   ls_output <- list(data = list(matrix = NULL,
@@ -707,7 +1009,7 @@ pheatmap_with_anno_combo <- function(
   
   # dendrogram
   cluster_condition <- !any(is.na(t_mat_cvd)) && !any(is.infinite(t_mat_cvd)) && 
-    NROW(t_mat_cvd) * NCOL(t_mat_cvd) <= 200 # gDR standard
+    any(dim(t_mat_cvd) < 200)  # gDR standard
   cluster_rows <- if (cluster_rows && cluster_condition && NROW(t_mat_cvd) >= 2) {
     stats::hclust(stats::dist(t_mat_cvd))
   } else {