+++ gDRplots/NEWS.md
@@ -1,6 +1,10 @@
-## gDRplots 0.0.35 - 2024-09-17
+## gDRplots 0.0.36 - 2024-09-25
 * add plots and help functions for PRISM data 
 
+## gDRplots 0.0.35 - 2024-09-23
+* update clustering condition
+* add `pheatmap_with_anno_cd` function
+ 
 ## gDRplots 0.0.34 - 2024-09-16
 * add clustering in `pheatmap_with_anno_sa` and `pheatmap_with_anno_combo` functions